Split puffin-cache into Puffin-specific and generic utilities
