+++ crates/cache-key/src/lib.rs
@@ -1,0 +1,8 @@
+pub use canonical_url::{CanonicalUrl, RepositoryUrl};
+pub use digest::digest;
+pub use stable_hash::{StableHash, StableHasher};
+
+mod cache_key;
+mod canonical_url;
+mod digest;
+mod stable_hash;
