+++ crates/cache-key/src/stable_hash.rs
